working; need to move items from pointcloud to GIS
