+++ public/js/IntegrationCode.js
@@ -5,8 +5,9 @@ import Zip from "../api-web-devs/tools/zip.js"
 import Evented from '../api-web-devs/components/evented.js'
 import oSettings from '../api-web-devs/components/settings.js'
 import PointCloudExtension from './pointcloud.js'
+import Loader from './loading.js'
 
-class IntegrationCode extends Autodesk.Viewing.Extension{
+export default class IntegrationCode extends Autodesk.Viewing.Extension{
 
     constructor(viewer, options){
         super(viewer, options);
@@ -16,7 +17,8 @@ class IntegrationCode extends Autodesk.Viewing.Extension{
 
     load(){
         console.log("Integration Code Extension is loaded");
-
+        var loader = new Loader();
+        loader.start();
         
         // var requests = [];
 
@@ -35,6 +37,8 @@ class IntegrationCode extends Autodesk.Viewing.Extension{
                 this.viewer.toolbar.removeControl(this._group);
             }
         }
+        console.log('IntegrationCode has been unloaded');
+        return true;
     }
 
     onToolbarCreated(){
@@ -237,114 +241,7 @@ class IntegrationCode extends Autodesk.Viewing.Extension{
         emit.Emit("error", { error:error })
 		//this.Emit("error", { error:error });
 	}
+  
 }
 
-    Autodesk.Viewing.theExtensionManager.registerExtension('IntegrationCode', IntegrationCode);
-
-
-    
-(function() {
-    class GisTool extends Autodesk.Viewing.ToolInterface {
-      constructor() {
-          super();
-          this.names = ['gis-tool'];
-   
-          // Hack: delete functions defined *on the instance* of the tool.
-          // We want the tool controller to call our class methods instead.
-          delete this.register;
-          delete this.deregister;
-          delete this.activate;
-          delete this.deactivate;
-          delete this.getPriority;
-          delete this.handleSingleClick;
-      }
-   
-      register() {      
-        console.log('GisTool registered.');
-      }
-   
-      deregister() {
-        this.viewer.unloadExtension('Autodesk.Geolocation');
-        this.geoTool = null;
-        console.log('GisTool unregistered.');
-      }
-   
-      async activate(name, viewer) {
-        this.viewer = viewer;
-        this.geoTool = await this.viewer.loadExtension('Autodesk.Geolocation');
-  
-        if (!this.geoTool.hasGeolocationData())
-          alert( 'No GIS data found in current model' );
-  
-        console.log('GisTool activated.');
-      }
-   
-      deactivate(name) {
-        console.log('GisTool deactivated.');
-      }
-   
-      getPriority() {
-        return 1; // Or feel free to use any number higher than 0 (which is the priority of all the default viewer tools)
-      }
-   
-      handleSingleClick(event, button) {
-        if (button === 0 ) {
-          // const hitPoint = this._intersect(event.clientX, event.clientY);
-          // const geolocation = this.geoTool.lmvToLonLat(hitPoint);
-          const canvasX = event.canvasX;
-          const canvasY = event.canvasY;
-          const res = this.viewer.clientToWorld(canvasX, canvasY);
-          const geolocation = this.geoTool.lmvToLonLat(res.point);
-          console.log(JSON.stringify(res.point));
-          console.log(JSON.stringify(geolocation)); //!<<< the geolocation you want
-          return true; // Stop the event from going to other tools in the stack
-        }
-        // Otherwise let another tool handle the event
-        return false;
-      }
-  
-      _intersect(clientX, clientY) {
-        return this.viewer.impl.intersectGround(clientX, clientY);
-      }
-    }
-  
-    class GisToolExtension extends Autodesk.Viewing.Extension {
-      constructor(viewer, options) {
-        super(viewer, options);
-        this.tool = new GisTool();
-      }
-   
-      load() {
-        this.viewer.toolController.registerTool(this.tool);
-        console.log('GisToolExtension loaded.');
-        return true;
-      }
-   
-      unload() {
-        this.viewer.toolController.deregisterTool(this.tool);
-        console.log('GisToolExtension unloaded.');
-        return true;
-      }
-   
-      onToolbarCreated(toolbar) {
-        const controller = this.viewer.toolController;
-        this.button = new Autodesk.Viewing.UI.Button('gis-tool-button');
-        this.button.onClick = (ev) => {
-            const isActivated = controller.isToolActivated('gis-tool');
-            if (isActivated) {
-              controller.deactivateTool('gis-tool');
-              this.button.setState(Autodesk.Viewing.UI.Button.State.INACTIVE);
-            } else {
-                controller.activateTool('gis-tool');
-                this.button.setState(Autodesk.Viewing.UI.Button.State.ACTIVE);
-            }
-        };
-        this.button.setToolTip('GIS Tool');
-  
-        this.group = new Autodesk.Viewing.UI.ControlGroup('gis-tool-group');
-        this.group.addControl(this.button);
-        toolbar.addControl(this.group);
-      }
-    }
-    Autodesk.Viewing.theExtensionManager.registerExtension('GisToolExtension', GisToolExtension);
-  })();
+Autodesk.Viewing.theExtensionManager.registerExtension('IntegrationCode', IntegrationCode);
