+++ public/js/pointcloud.js
@@ -1,4 +1,5 @@
-import Loader from './loading.js'
+import { GisTool } from './GISintegration.js';
+// import Loader from './loading.js'
 
 let geometry;
 
@@ -7,6 +8,9 @@ export default class PointCloudExtension extends Autodesk.Viewing.Extension {
         super(viewer, options);
 		this._group = null;
         this._button = null;
+
+        //gis added
+        this.tool = new GisTool();
 
         this.zaxisOffset = -1.5;
 
@@ -22,14 +26,22 @@ export default class PointCloudExtension extends Autodesk.Viewing.Extension {
     }
 
     load() {
-        //window.renderPointCloud = false;
-        console.log('Point Cloud extension loaded')
-        var loader = new Loader();
+        // window.renderPointCloud = false;
+        // console.log('Point Cloud extension loaded')
+        // var loader = new Loader(); //this loads the button
         // var simulationData = 
-        loader.start();
-        //console.log(test.simulationObj);
-        //this._renderCloud(simulationData);
+        // loader.start();
+        // console.log(test.simulationObj);
+        // this._renderCloud(simulationData);
+        // return true;
+
+        //gis added
+        this.viewer.toolController.registerTool(this.tool);
+        console.log('GisToolExtension loaded.');
+        // var loader = new Loader();
+        // loader.start
         return true;
+        
     }
 
     _renderCloud(){
@@ -43,14 +55,25 @@ export default class PointCloudExtension extends Autodesk.Viewing.Extension {
     }
 
     unload() {
-		  // Clean our UI elements if we added any
-         if (this._group) {
+		//   // Clean our UI elements if we added any
+        //  if (this._group) {
+        //     this._group.removeControl(this._button);
+        //     if (this._group.getNumberOfControls() === 0) {
+        //         this.viewer.toolbar.removeControl(this._group);
+        //     }
+        // }
+        // console.log('PointCloudExtension has been unloaded');
+        // return true;
+
+        //gis added
+        this.viewer.toolController.deregisterTool(this.tool);
+        if (this._group) {
             this._group.removeControl(this._button);
             if (this._group.getNumberOfControls() === 0) {
                 this.viewer.toolbar.removeControl(this._group);
             }
         }
-        console.log('PointCloudExtension has been unloaded');
+        console.log('GisToolExtension unloaded.');
         return true;
     }
     
@@ -137,31 +160,66 @@ export default class PointCloudExtension extends Autodesk.Viewing.Extension {
 	
 	onToolbarCreated() {
         // Create a new toolbar group if it doesn't exist
+        // this._group = this.viewer.toolbar.getControl('allMyAwesomeExtensionsToolbar');
+        // if (!this._group) {
+        //     this._group = new Autodesk.Viewing.UI.ControlGroup('allMyAwesomeExtensionsToolbar');
+        //     this.viewer.toolbar.addControl(this._group);
+        // }
+
+        // // Add a new button to the toolbar group
+        // this._button = new Autodesk.Viewing.UI.Button('myAwesomeExtensionButton');
+        // this._button.onClick = (ev) => {
+            
+        //     window.renderPointCloud = true;
+        //     this._renderCloud();
+        //     if(window.legendOFF){
+        //         this._appearLegend();
+        //         window.legendOFF = false;
+        //     }
+
+        //     var i = 0;
+                
+        //     var interval = setInterval(() => {
+        //         this._updatePointCloudGeometry(simulation.Frames[i]);
+        //         if (++i == simulation.Frames.length) window.clearInterval(interval);
+        //     }, 50);
+        // };
+
+        // this._button.setToolTip('Sim Object');
+        // this._button.addClass('connectCadmiumIcon');
+        // this._group.addControl(this._button);
+
+        //gis added
+        // Create a new toolbar group if it doesn't exist
         this._group = this.viewer.toolbar.getControl('allMyAwesomeExtensionsToolbar');
         if (!this._group) {
             this._group = new Autodesk.Viewing.UI.ControlGroup('allMyAwesomeExtensionsToolbar');
             this.viewer.toolbar.addControl(this._group);
         }
 
+        const controller = this.viewer.toolController;
+
         // Add a new button to the toolbar group
         this._button = new Autodesk.Viewing.UI.Button('myAwesomeExtensionButton');
         this._button.onClick = (ev) => {
-            //window.renderPointCloud = true;
-            this._renderCloud();
-            if(window.legendOFF){
-                this._appearLegend();
-                window.legendOFF = false;
+            const isActivated = controller.isToolActivated('gis-tool');
+            if(isActivated){
+                controller.deactivateTool('gis-tool');
+                this._button.setState(Autodesk.Viewing.UI.Button.State.INACTIVE);
+            } else {
+                controller.activateTool('gis-tool');
+                this._button.setState(Autodesk.Viewing.UI.Button.State.ACTIVE);
             }
-
-            var i = 0;
-                
-            var interval = setInterval(() => {
-                this._updatePointCloudGeometry(simulation.Frames[i]);
-                if (++i == simulation.Frames.length) window.clearInterval(interval);
-            }, 50);
+            
+            // window.renderPointCloud = true;
+            // this._renderCloud();
+            // if(window.legendOFF){
+            //     this._appearLegend();
+            //     window.legendOFF = false;
+            // }
         };
 
-        this._button.setToolTip('Simulation Object');
+        this._button.setToolTip('GIS Tool');
         this._button.addClass('connectCadmiumIcon');
         this._group.addControl(this._button);
     }
@@ -292,6 +350,5 @@ export default class PointCloudExtension extends Autodesk.Viewing.Extension {
                     .call(axisLeg);
     }
 }
-
 
 Autodesk.Viewing.theExtensionManager.registerExtension('PointCloudExtension', PointCloudExtension);
